+++ SC_Synthetic_UH_Method.py
@@ -7,7 +7,7 @@ from Rainfall_Data_Curves import rainfall_data_curves
 
 
 # Combines rainfallDistributionCurve, PRFData, weightedCurveNumber, and travelTimeMethodTimeOfConcentration or lagTimeMethodTimeOfConcentration (depending on TcMethod) into single function.
-def calculateMissingParametersSCSUH(lat, lon, AEP, curveNumberMethod, TcMethod, length=None, slope=None, dataSheetFlow=None, dataExcessSheetFlow=None, dataShallowConcentratedFlow=None, dataChannelizedFlowOpenChannel=None, dataChannelizedFlowStormSewer=None, dataChannelizedFlowStormSewerOrOpenChannelUserInputVelocity=None):
+def calculateMissingParametersSCSUH(lat, lon, prfData, AEP, curveNumberMethod, TcMethod, length=None, slope=None, dataSheetFlow=None, dataExcessSheetFlow=None, dataShallowConcentratedFlow=None, dataChannelizedFlowOpenChannel=None, dataChannelizedFlowStormSewer=None, dataChannelizedFlowStormSewerOrOpenChannelUserInputVelocity=None):
     # AEP: 10 - 10 year return period, 4 - 25 year return period, 2 - 50 year return period, 1 - 100 year return period
     # curveNumberMethod: "runoff" or "area"
     # TcMethod: 'lagtime' or 'traveltime'
@@ -36,14 +36,14 @@ def calculateMissingParametersSCSUH(lat, lon, AEP, curveNumberMethod, TcMethod, 
                                         dataChannelizedFlowOpenChannel,
                                         dataChannelizedFlowStormSewer,
                                         dataChannelizedFlowStormSewerOrOpenChannelUserInputVelocity)
-            Tc = {'value': Tc[0],'warningMessage': Tc[1] }
+            Tc = {'value': Tc }
         else:
             raise Exception("Not all parameters for traveltime were entered.")
     elif TcMethod.lower() == "lagtime":
         if all([length, slope]):
             rainfall_distribution_curve_number = rainfall_distribution_curve[1]
             Tc = lagTimeMethodTimeOfConcentration(length, slope, rainfall_distribution_curve_number)
-            Tc = {'value': Tc,'warningMessage': None }
+            Tc = {'value': Tc }
 
         else:
             raise Exception("Not all parameters for lagtime were entered.")
@@ -51,7 +51,7 @@ def calculateMissingParametersSCSUH(lat, lon, AEP, curveNumberMethod, TcMethod, 
         raise Exception("Time of concentration method not valid.")
     
     # Get PRF
-    PRF = PRFData (lat, lon)
+    PRF = PRFData (prfData)
     
     # Get Curve Number, S, Ia
     if curveNumberMethod.lower() == "runoff" or curveNumberMethod.lower() == "area":
@@ -151,25 +151,12 @@ def areaWeightedCN(curveNumberData, P24hr):
 
 # Extracts data from PRF GIS layer 
 # Corresponds to "PRF Calculator" sheet in spreadsheet
-def PRFData(lat, lon):
-    
-
-    # Placeholder data-- waiting on GIS data to be published
-    PRFData = [
-        {
-            "PRF": 180,
-            "Area": 50.0
-        },
-        {
-            "PRF": 300,
-            "Area": 50.0
-        }
-    ]
+def PRFData(prfData):
 
     # Calculate Watershed PRF and Shape Parameter, n (also called Gamma_n)
     total_watershed_area = 0.0
     total_product = 0.0
-    for row in PRFData:
+    for row in prfData:
         total_watershed_area += row["Area"]
         total_product += row["PRF"] * row["Area"]
     PRF = total_product / total_watershed_area